% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bz-family.R
\name{compute_bz_index}
\alias{compute_bz_index}
\title{compute_bz_index}
\usage{
compute_bz_index(wvg)
}
\arguments{
\item{wvg}{a weigthed or simple voting game.}
}
\value{
A vector of Bz indexes The order corresponds to the order of the voters.
}
\description{
Computes the Bz index for each voter. The Bz index is the Bz score normalised.
}
\details{
The current algorithm has an exponential running time in terms of the
number of voters.
}

