% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bz-family.R
\name{compute_bz_power}
\alias{compute_bz_power}
\title{compute_bz_power}
\usage{
compute_bz_power(wvg)
}
\arguments{
\item{wvg}{a weigthed or simple voting game.}
}
\value{
A vector of Bz power measures. The order corresponds to the order
of the voters.
}
\description{
The Banthaf voting power measure is corresponds to the probability that
a voter is critical.
}
\details{
The current algorithm has an exponential running time in terms of the
number of voters.
}

