% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bz-family.R
\name{compute_bz_score}
\alias{compute_bz_score}
\title{compute_bz_score}
\usage{
compute_bz_score(wvg)
}
\arguments{
\item{wvg}{a weigthed or simple voting game.}
}
\value{
A vector of Bz scores. The order corresponds to the order of the voters.
}
\description{
Computes the Bz score for each voter. The Bz score is the number of coalitions
where that voter is critical; i.e. where voting no changes the outcome.
}
\details{
The current algorithm has an exponential running time in terms of the
number of voters.
}

