% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-structures.R
\name{create_equal_weight_voting_game}
\alias{create_equal_weight_voting_game}
\title{create_equal_weight_voting_game}
\usage{
create_equal_weight_voting_game(quota, n)
}
\arguments{
\item{quota}{the quota of the voting game.}

\item{n}{the number of voters}
}
\value{
A list datastructure with two values: quota and weights
corresponding to the input values.
}
\description{
A voting game with an assembly with a quota and a number of voters
with equal weight of one.
}

