% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data-structures.R
\name{create_weighted_voting_game}
\alias{create_weighted_voting_game}
\title{create_weighted_voting_game}
\usage{
create_weighted_voting_game(quota, weights)
}
\arguments{
\item{quota}{the quota of the voting game.}

\item{weights}{A vector of weights. Each weight corresponds to one voter.}
}
\value{
A list datastructure with two values: quota and weights
corresponding to the input values.
}
\description{
A weigthed voting game consists of a quota and an assembly.
In this case the quota is the number of weigthed votes that are
needed to change the outcome to yes.
Each voter has a weight.
}

