% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/votingpower.R
\docType{package}
\name{votingpower}
\alias{votingpower}
\alias{votingpower-package}
\title{votingpower}
\description{
Implements voting power indexes and algorithms.
}
\examples{
# The famous Nassau County Board's voting system
# [https://en.wikipedia.org/wiki/Banzhaf_power_index]
board <- create_weighted_voting_game(16, c(9, 9, 7, 3, 1, 1))
compute_bz_power(board)
}

